requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.1
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.21
alembic>=1.12.0
asyncpg>=0.28.0
supabase>=1.2.0
aiosqlite>=0.20.0
email-validator>=2.0.0